segment tests remove original (now split) billing invoice generation tests file updates to our taxation tests feat: distinguish between discounts and credits in tax calculation logic
This commit updates our tax calculation documentation to clearly distinguish between two types of negative line items:

1. Discount items: Promotional reductions that do NOT affect the taxable base
2. Credit items: Actual reductions to services that DO reduce the taxable base

The update provides more detailed explanations of how each item type impacts tax calculations, adds a new example scenario demonstrating credit handling, and restructures the document for better clarity. This distinction is critical for proper tax compliance in various jurisdictions.
